6DOF net update
