+++ src/6DOF_void.cpp
@@ -51,10 +51,7 @@ void sixdof_void::ini(lexer *p, ghostcell *pgc)
 
 void sixdof_void::start_cfd(lexer* p, fdm* a, ghostcell* pgc, vrans* pvrans, vector<net*>& pnet, int iter, field &uvel, field &vvel, field &wvel, field &fx, field &fy, field &fz, bool finalize)
 {
-    
-    //if(p->mpirank==0)
-    //cout<<"6DOF: 001"<<endl;
-    
+
     if (p->X310 > 0)
     {
         for (int i=0; i<p->mooring_count; i++)
@@ -63,26 +60,16 @@ void sixdof_void::start_cfd(lexer* p, fdm* a, ghostcell* pgc, vrans* pvrans, vec
         }
     }
     
-    //if(p->mpirank==0)
-    //cout<<"6DOF: 002"<<endl;
-    
     if (p->X320 > 0)
     {
         for (int ii = 0; ii < p->net_count; ii++)
         {
             pnet[ii]->start(p, a, pgc, 1.0, quatRotMat);
-            if(p->mpirank==0)
-            cout<<"6DOF: 003"<<endl;
             pvrans->start(p, a, pgc, pnet[ii], ii);
-            if(p->mpirank==0)
-            cout<<"6DOF: 004"<<endl;
 
             // Forces on rigid body
             pnet[ii]->netForces(p,Xne[ii],Yne[ii],Zne[ii],Kne[ii],Mne[ii],Nne[ii]);
-            
-            if(p->mpirank==0)
-            cout<<"6DOF: 005"<<endl;
-        
+
             if( p->mpirank == 0)
             {
                 cout<<"Xne"<< ii <<" : "<<Xne[ii]<<" Yne"<< ii <<" : "<<Yne[ii]<<" Zne"<< ii <<" : "<<Zne[ii]
@@ -91,225 +78,9 @@ void sixdof_void::start_cfd(lexer* p, fdm* a, ghostcell* pgc, vrans* pvrans, vec
         }
     }
     
-    //if(p->mpirank==0)
-    //cout<<"6DOF: 006"<<endl;
-    
-    
     ++p->printcount_sixdof;
 }
 
-void sixdof_void::initialize(lexer *p, fdm *a, ghostcell *pgc, vector<net*>& pnet)
-{
-    //if(p->mpirank==0)
-    //cout<<"6DOF: ini"<<endl;
-    
-    
-	if(p->X310==0)
-	{
-		pmooring.push_back(new mooring_void());
-	}
-	else
-	{
-		MPI_Bcast(&p->mooring_count,1,MPI_DOUBLE,0,pgc->mpi_comm);	
-        
-        Xme.resize(p->mooring_count);
-		Yme.resize(p->mooring_count);
-		Zme.resize(p->mooring_count);
-		Kme.resize(p->mooring_count);
-		Mme.resize(p->mooring_count);
-		Nme.resize(p->mooring_count);
-
-		if(p->mpirank==0)
-		{
-			mkdir("./REEF3D_CFD_6DOF_Mooring",0777);	
-		}		
-
-		pmooring.reserve(p->mooring_count);
-			
-		for (int i=0; i < p->mooring_count; i++)
-		{
-			if(p->X310==1)
-			{
-				pmooring.push_back(new mooring_Catenary(i));
-			}	
-			else if(p->X310==2)
-			{
-				pmooring.push_back(new mooring_barQuasiStatic(i)); 
-			}	
-			else if(p->X310==3)
-			{
-				pmooring.push_back(new mooring_dynamic(i));
-			}
-			else if(p->X310==4)
-			{
-                pmooring.push_back(new mooring_Spring(i));
-			}
-		
-			pmooring[i]->initialize(p,pgc);
-		}
-	}
-
-    if (p->X320 == 0)
-    {
-        pnet.push_back(new net_void());
-    }
-    
-    else
-    {
-		MPI_Bcast(&p->net_count,1,MPI_DOUBLE,0,pgc->mpi_comm);
-        
-        Xne.resize(p->net_count);
-		Yne.resize(p->net_count);
-		Zne.resize(p->net_count);
-		Kne.resize(p->net_count);
-		Mne.resize(p->net_count);
-		Nne.resize(p->net_count);
-
-        if(p->mpirank==0)
-        mkdir("./REEF3D_CFD_6DOF_Net",0777);	
-
-        else
-        {
-            p->X320_type = new int[p->net_count];
-        }
-		
-        pnet.reserve(p->net_count);	
-  
-		for (int ii=0; ii < p->net_count; ii++)
-		{
-            MPI_Bcast(&p->X320_type[ii],1,MPI_INT,0,pgc->mpi_comm);
-			
-            if(p->X320_type[ii] > 10)
-			{
-				pnet.push_back(new net_barDyn(ii,p));
-			}
-            else if (p->X320_type[ii] < 4)
-            {
-                pnet.push_back(new net_barQuasiStatic(ii,p));
-            }
-            else
-            {
-                 pnet.push_back(new net_sheet(ii,p));
-            }
-			
-            pnet[ii]->initialize(p,a,pgc);
-		}
-    }
-
-    // Ini parameters
-	p->ufbi=p->vfbi=p->wfbi=0.0;
-	p->pfbi=p->qfbi=p->rfbi=0.0;
-    p->xg=p->yg=p->zg=0.0;
-
-    quatRotMat << 1.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 1.0;
-}
-
-
-void sixdof_void::initialize(lexer *p, fdm_nhf *d, ghostcell *pgc, vector<net*>& pnet)
-{
-    //if(p->mpirank==0)
-    //cout<<"6DOF: ini"<<endl;
-    
-    /*
-	if(p->X310==0)
-	{
-		pmooring.push_back(new mooring_void());
-	}
-	else
-	{
-		MPI_Bcast(&p->mooring_count,1,MPI_DOUBLE,0,pgc->mpi_comm);	
-        
-        Xme.resize(p->mooring_count);
-		Yme.resize(p->mooring_count);
-		Zme.resize(p->mooring_count);
-		Kme.resize(p->mooring_count);
-		Mme.resize(p->mooring_count);
-		Nme.resize(p->mooring_count);
-
-		if(p->mpirank==0)
-		{
-			mkdir("./REEF3D_CFD_6DOF_Mooring",0777);	
-		}		
-
-		pmooring.reserve(p->mooring_count);
-			
-		for (int i=0; i < p->mooring_count; i++)
-		{
-			if(p->X310==1)
-			{
-				pmooring.push_back(new mooring_Catenary(i));
-			}	
-			else if(p->X310==2)
-			{
-				pmooring.push_back(new mooring_barQuasiStatic(i)); 
-			}	
-			else if(p->X310==3)
-			{
-				pmooring.push_back(new mooring_dynamic(i));
-			}
-			else if(p->X310==4)
-			{
-                pmooring.push_back(new mooring_Spring(i));
-			}
-		
-			pmooring[i]->initialize(p,pgc);
-		}
-	}
-
-    if (p->X320 == 0)
-    {
-        pnet.push_back(new net_void());
-    }
-    
-    else
-    {
-		MPI_Bcast(&p->net_count,1,MPI_DOUBLE,0,pgc->mpi_comm);
-        
-        Xne.resize(p->net_count);
-		Yne.resize(p->net_count);
-		Zne.resize(p->net_count);
-		Kne.resize(p->net_count);
-		Mne.resize(p->net_count);
-		Nne.resize(p->net_count);
-
-        if(p->mpirank==0)
-        mkdir("./REEF3D_CFD_6DOF_Net",0777);	
-
-        else
-        {
-            p->X320_type = new int[p->net_count];
-        }
-		
-        pnet.reserve(p->net_count);	
-  
-		for (int ii=0; ii < p->net_count; ii++)
-		{
-            MPI_Bcast(&p->X320_type[ii],1,MPI_INT,0,pgc->mpi_comm);
-			
-            if(p->X320_type[ii] > 10)
-			{
-				pnet.push_back(new net_barDyn(ii,p));
-			}
-            else if (p->X320_type[ii] < 4)
-            {
-                pnet.push_back(new net_barQuasiStatic(ii,p));
-            }
-            else
-            {
-                 pnet.push_back(new net_sheet(ii,p));
-            }
-			
-            pnet[ii]->initialize(p,a,pgc);
-		}
-    }*/
-
-    // Ini parameters
-	p->ufbi=p->vfbi=p->wfbi=0.0;
-	p->pfbi=p->qfbi=p->rfbi=0.0;
-    p->xg=p->yg=p->zg=0.0;
-
-    quatRotMat << 1.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 1.0;
-}
 
 void sixdof_void::start_nhflow(lexer* p, fdm_nhf* d, ghostcell* pgc, vrans* pvrans, vector<net*>& pnet, int iter, 
                                         double *U, double *V, double *W, double *FX, double *FY, double *FZ, slice &WL, slice &fe, bool finalize)
